Fix inaccuracy because of missing timer resets
Reset counter after each check to prevent counting the overhead of
the redis updates. When a bulk update happend the elapsed time was
added to the time of the check.

Also reset counter for skipped rules. This caused that the time of
the skipped rules was added to the next rule.